Create utils to media (canUpload)
